possible accidental rule settings change
due to the fact that canceling  the settings menu set your value to 1, which caused the value to tick up every time you closed it, leading to a possible setting change later down the line. Added two prevention methods. Setting score to 0 every time the decode manager runs, and fails the checksum, as well as making the cancel button set the menu score to 0